# Conflicts:
#	client-app/src/components/form/InputToggle.js
#	config/app.php
#	customer-portal-app/build/asset-manifest.json
#	customer-portal-app/build/index.html
#	customer-portal-app/build/precache-manifest.83e82ee24a90daa315b52152b5c9c62c.js
#	customer-portal-app/build/service-worker.js
#	resources/views/welcome.blade.php

+++ app/Http/Controllers/Api/Project/RevenuesKwhController.php
@@ -315,16 +315,22 @@ class RevenuesKwhController extends ApiController
             . $document->filename));
         file_put_contents($filePath, $pdf);
 
-        $alfrescoHelper = new AlfrescoHelper(\Config::get('app.ALFRESCO_COOP_USERNAME'), \Config::get('app.ALFRESCO_COOP_PASSWORD'));
+        if(\Config::get('app.ALFRESCO_COOP_USERNAME') != 'local') {
+            $alfrescoHelper = new AlfrescoHelper(\Config::get('app.ALFRESCO_COOP_USERNAME'), \Config::get('app.ALFRESCO_COOP_PASSWORD'));
 
-        $alfrescoResponse = $alfrescoHelper->createFile($filePath,
-            $document->filename, $document->getDocumentGroup()->name);
+            $alfrescoResponse = $alfrescoHelper->createFile($filePath,
+                $document->filename, $document->getDocumentGroup()->name);
 
-        $document->alfresco_node_id = $alfrescoResponse['entry']['id'];
+            $document->alfresco_node_id = $alfrescoResponse['entry']['id'];
+        }else{
+            $document->alfresco_node_id = null;
+        }
         $document->save();
 
         //delete file on server, still saved on alfresco.
-        Storage::disk('documents')->delete($document->filename);
+        if(\Config::get('app.ALFRESCO_COOP_USERNAME') != 'local') {
+            Storage::disk('documents')->delete($document->filename);
+        }
     }
 
     public function destroy(RevenuesKwh $revenuesKwh)
@@ -436,8 +442,10 @@ class RevenuesKwhController extends ApiController
 
             return $pdf;
         } else {
+            $revenueHtml = '<h1>Kon geen document opstellen: adres niet compleet</h1>';
+
             $pdf = PDF::loadView('documents.generic', [
-                'html' => '',
+                'html' => $revenueHtml,
             ])->output();
 
             return $pdf;
@@ -477,7 +485,7 @@ class RevenuesKwhController extends ApiController
 
             //Make preview email
             if ($primaryEmailAddress) {
-                $mailbox = $this->setMailConfigByDistribution($distributionKwh);
+                $mailbox = $this->setMailConfigByDistribution($project);
                 if ($mailbox) {
                     $fromEmail = $mailbox->email;
                     $fromName = $mailbox->name;
@@ -688,28 +696,33 @@ class RevenuesKwhController extends ApiController
                     . 'documents/' . $document->filename));
                 file_put_contents($filePath, $pdf);
 
-                $alfrescoHelper = new AlfrescoHelper(\Config::get('app.ALFRESCO_COOP_USERNAME'),
-                    \Config::get('app.ALFRESCO_COOP_PASSWORD'));
+                if(\Config::get('app.ALFRESCO_COOP_USERNAME') != 'local') {
+                    $alfrescoHelper = new AlfrescoHelper(\Config::get('app.ALFRESCO_COOP_USERNAME'),
+                        \Config::get('app.ALFRESCO_COOP_PASSWORD'));
 
-                $alfrescoResponse = $alfrescoHelper->createFile($filePath,
-                    $document->filename, $document->getDocumentGroup()->name);
-                if($alfrescoResponse == null)
-                {
-                    throw new \Exception('Fout bij maken rapport document in Alfresco.');
+                    $alfrescoResponse = $alfrescoHelper->createFile($filePath,
+                        $document->filename, $document->getDocumentGroup()->name);
+                    if($alfrescoResponse == null)
+                    {
+                        throw new \Exception('Fout bij maken rapport document in Alfresco.');
+                    }
+                    $document->alfresco_node_id = $alfrescoResponse['entry']['id'];
+                }else{
+                    $document->alfresco_node_id = null;
                 }
-                $document->alfresco_node_id = $alfrescoResponse['entry']['id'];
+
                 $document->save();
             }
             catch (\Exception $e) {
                 Log::error('Fout bij maken rapport document voor ' . ($primaryEmailAddress ? $primaryEmailAddress->email : '**onbekend emailadres**') . ' (' . $contact->full_name . ')' );
                 Log::error($e->getMessage());
-                array_push($messages, 'Fout bij maken rapport document voor ' . $primaryEmailAddress->email . ' (' . $contact->full_name . ')' );
+                array_push($messages, 'Fout bij maken rapport document voor ' . ($primaryEmailAddress ? $primaryEmailAddress->email : '**onbekend emailadres**') . ' (' . $contact->full_name . ')' );
             }
 
             //send email
             if ($primaryEmailAddress) {
                 try{
-                    $mailbox = $this->setMailConfigByDistribution($distributionKwh);
+                    $mailbox = $this->setMailConfigByDistribution($project);
                     if ($mailbox) {
                         $fromEmail = $mailbox->email;
                         $fromName = $mailbox->name;
@@ -777,16 +790,14 @@ class RevenuesKwhController extends ApiController
                 }
 
             } else {
-                return [
-                    'from' => 'Geen e-mail bekend.',
-                    'to' => 'Geen e-mail bekend.',
-                    'subject' => 'Geen e-mail bekend.',
-                    'htmlBody' => 'Geen e-mail bekend.'
-                ];
+                Log::error( 'Fout bij verzenden email naar **onbekend emailadres** (' . $contact->full_name . ')' );
+                array_push($messages, 'Fout bij verzenden email naar **onbekend emailadres** (' . $contact->full_name . ')' );
             }
 
             //delete file on server, still saved on alfresco.
-            Storage::disk('documents')->delete($document->filename);
+            if(\Config::get('app.ALFRESCO_COOP_USERNAME') != 'local') {
+                Storage::disk('documents')->delete($document->filename);
+            }
         }
         if(count($messages) > 0)
         {
@@ -798,13 +809,10 @@ class RevenuesKwhController extends ApiController
         }
     }
 
-    protected function setMailConfigByDistribution(RevenueDistributionKwh $distributionKwh)
+    protected function setMailConfigByDistribution($project)
     {
         // Standaard vanuit primaire mailbox mailen
         $mailboxToSendFrom = Mailbox::getDefault();
-
-        // Als er een mailbox aan de administratie is gekoppeld, dan deze gebruiken
-        $project = $distributionKwh->revenuesKwh->project;
 
         if ($project->administration && $project->administration->mailbox) {
             $mailboxToSendFrom = $project->administration->mailbox;
@@ -820,11 +828,10 @@ class RevenuesKwhController extends ApiController
     protected function translateToValidCharacterSet($field){
 
         $field = strtr(utf8_decode($field), utf8_decode('ÀÁÂÃÄÅÆÇÈÉÊËÌÍÎÏÐÑÒÓÔÕÖØÙÚÛÜÝßàáâãäåæçèéêëìíîïðñòóôõöøùúûüýÿ'), 'AAAAAAACEEEEIIIIDNOOOOOOUUUUYsaaaaaaaceeeeiiiionoooooouuuuyy');
-        $field = iconv('UTF-8', 'ASCII//TRANSLIT', $field);
+//        $field = iconv('UTF-8', 'ASCII//TRANSLIT', $field);
         $field = preg_replace('/[^A-Za-z0-9 -]/', '', $field);
 
         return $field;
     }
-
 
 }
